test app with graphics
figured out how to make the graphics work in a simple demo app
apart from that simple objects created to automate the workflow. To this
end several classes have been cretated. See details

Objects/Classes are scoped using VBF namespace (VBF stands for vibratroy
        bowl feeder) to avoid conflicts with bullets. Apart from that
several typedef have been introduced to get rid of ugly long names in
bullet physics library. The rule of thumb is that objects/classes are
scoped in VBF namespace and typedefs are not scoped (till this point)
which means some object not scoped with VBF will refer directly to a
bullet object while one scoped with VBF will refer to an object created
specifically for this project

1. VBF::World : as used in bullet terminology, the physics world
2. VBF::RigidBody : object that creates a rigid bodies
3. VBF::CommonPhysics : object that takes world, rigid bodies and makes
the physics environment out of it
4. VBF::Sphere and VBF::Cube : objects that creates simple sphere and
cubes