Cleanup and initial documentation
Reasons for deleting files
--------------------------
I feel the project has reached a stage where I don't need to look at
demo codes for my understanding. Hence I deleted the old Demo files from
repository. However they still exist in the previous commits of this
repository and can be referred to if needed.

Initial Documentation
---------------------
The doxygen config file called 'dconfig' has been added to the
repository and this will be used henceforth to generate the
doxygen documentation. Doxygen and Latex must be installed on
the host system for generating doxygen documentation