Unit Tests with Google Tests
Framework for unit tests built into the project.
The testing suite for this project is GoogleTest
which in the latest version includes both gmock and gtest

TDD should be the preferred way to develop the project from this point
onwards.

TDD can serve as a nice documentation of what specifications the project
comply to.

TODO: Figure out a test for already developed codes.