bug in input file detected and corrected
The bug for reading data from input file has been corrected and removed.
The input file data is now stored in an unordered hash map for.

The variable name is string and the value is assumed to be a float
literal that is associated with the string variable name and separated
from the string by a space character